+++ api/src/graphql/index.ts
@@ -1,0 +1,4 @@
+import itemsSchema from './items/items.graphql';
+import usersSchema from './users/users.graphql';
+
+export { itemsSchema, usersSchema };
